Keep script in same file. Attempting to fix esm
